refactor: extract deck start feature
