+++ src/pages/deck-start/ui/DeckStartView.stories.tsx
@@ -1,4 +1,4 @@
-import { DeckStartForm } from "@/features/study";
+import { DeckStartForm } from "@/features/deck-start";
 
 import type { Meta, StoryObj } from "@storybook/react";
 import type { ComponentProps } from "react";
